requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic>=2.0.0
loguru==0.7.2
tenacity==8.2.3
anthropic==0.49.0
google-cloud-texttospeech==2.19.0